Reorganize modules
Move everything related to the app into the app module.  This reduces
the number of top-level modules and separates app implementation code
(widgets, searchprovider, model, etc.) from generic support infra
(networking code, gettext, etc.)
