+++ src/app/model/device.rs
