+++ src/app/widgets/validation_indicator.rs
